refactor: disentangle logic from template rendering using Twig
